LayoutTests/requestidlecallback/requestidlecallback-does-not-block-timer-expected.txt

This tests that scheduling an idle callback would not delay the firing of a timer.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS didFireTimer is true
PASS idleDeadline.timeRemaining() > 0 is true
PASS idleDeadline.timeRemaining() is 0
PASS didRunIdleCallback is true
PASS successfullyParsed is true

TEST COMPLETE
